// src/adapters/infrastructure/frameworks/http/dtos/login_password.dto.ts
// -------------------------------------------------------
// Helpers
// -------------------------------------------------------

interface LoginPasswordDto {
  photoUrl: string;
  passwordPiece: string;
  photoRelativePath: string;
  photoName: string;
}

// -------------------------------------------------------
// Public Interface
// -------------------------------------------------------

export type { LoginPasswordDto };